Update the layout to be closer to the factory layout
Adds `Fn` and `Mouse` layers, so the switcher keys actually work. Also
rearranged `Mod` to follow the factory layout, with the added change that
`Shift` keys are also included on the layer. The `Fn` layer was laid out,
according to the factory layout. The `Mouse` layer is mostly empty, except for
some modifiers and the switcher key.

The most significant difference from the factory layout is that modifiers are
present on all layers, in the same position.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>