fixes for HID class requests
So far there was a mix of using the handle's variables as buffer (protocol),
and as actual value storage (idleRate). Let's decide to store the protocol
also in the handle instead of in separate variable. This requires the
proper reset of these values at each SET_CONFIGURATION. This also means
that the GET_IDLE and GET_PROTOCOL requests don't need to be processed by
the interfaces.

Add the missing success flags to the callback paths where the request is
accepted. Remove empty cases. Only call back if the descriptor indicates
BOOT subclass.