remove external report scancode indexing
This indexing offers little performance benefit at the cost of exposing
the state of the report buffer to an external variable, creating
the possibility of conflicts when only one of them are modified somewhere.
Since the NKRO bitfield implementation will make this indexing obsolete anyway,
get rid of it in advance.