move report descriptors inline into header
The purpose of this change is to be able to modify the report descriptors
inline, without having to separately adjust the hardcoded length.
By using static keyword, the descriptor will be copied into every translation unit,
but this isn't a problem even without smart compilers, because these
descriptors are only used once, even if their size is used in multiple places.