Fix double unschedule on callMacro.
callMacro manually unschedules the macro when putting it to sleep, and
returns as finished. If it is also the last command of a macro, the macro
gets unscheduled again because it has ended.

This would lead to unscheduleCurrentSlot() being called twice. Original
double-unschedule protection would not work, because
scheduler.currentSlotIdx gets automatically updated when unscheduling the
slot, therefore leading to another macro being unscheduled forever.