Separate x and y high-res scrolling and use values in SetReport requests
Allow high-res scrolling to be selectively enabled for the two axes, as
specified by the values given in the SetReport feature request. AFAIK
only Windows supports this (via registry keys[^1]).

Beware that disabling high-res scrolling for only one axis doesn't play
too well with axis locking. I've adjusted as best as I could, but
there's ultimately no getting around it: a movement with equal physical
x and y displacement will have much larger values on the scaled axis. To
try to adjust for this, there could be some kind of extra buffer that
must be filled before scroll values are sent, but this would induce a
delay before movement is seen on the high-res axis, disrupting the
smooth, 1:1 (or nearly 1:1) feel provided by the higher resolution,
defeating its purpose (to an extent).

I think the right way to deal with this would be to change the USB
descriptor for the mouse so that there's only one resolution multiplier.
Because of the way resolution multipliers are scoped, I think the
logical groups for the wheels would have to be removed so that a single
resolution multiplier in the descriptor can apply to both wheels. At
least, this is the case from my recollection of the USB HID spec. It's
possible that OS drivers would take a global resolution multiplier to
apply to any contained logical collections that don't themselves have a
resolution multiplier specified.

[^1]: [Enhanced Wheel Support in
Windows](https://learn.microsoft.com/en-us/previous-versions/windows/hardware/design/dn613912(v=vs.85))