Initialize `usbMouseFeatBuffer` globally
When waking from sleep, a Linux host will communicate with the device
such that `kUSB_DeviceEventSetConfiguration` is triggered, but won't
send `SetReport` like it does when first initializing the device upon
connection. This broke scrolling when waking from sleep since the
`usbMouseFeatBuffer` got re-initialized to `0` but never got the
`SetReport` that would set it back to `1`, resulting in the device
thinking it should operate in low-res scrolling mode while the host
continues to expect high-resolution scrolling.

To fix this, I've removed the initialization of `usbMouseFeatBuffer`
from `kUSB_DeviceEventSetConfiguration` and instead initialize it
globally, when it's declared.

Fixes #1261.